Clarified the Arrays.asList command
Geeky rationale:
Arrays.asList does not return java.util.ArrayList, it returns Arrays$ArrayList, which is a different beast. This could trip some people up, so let's keep it as simple as possible.
The ArrayList we all know and love is modifiable and this list is not.
If we check the docs for Arrays.asList, we see that it does not promise any particular list implementation, it just says List.